# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Compte API",
        "description": "API Documentation pour la gestion des comptes utilisateurs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Obtenir la liste de toutes les catégories",
                "operationId": "0a3810ef08509327a6db442d3e4c4b73",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Categorie"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categorieStore": {
            "post": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Créer une nouvelle catégorie",
                "operationId": "f683298625373815e746b4458d33febc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie ajoutée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categorieShow{id}": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Obtenir les détails d'une catégorie spécifique",
                "operationId": "6a58091d9e5084a2fa99222878a57cd7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categorie": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Categorie"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categorieUpdate{id}": {
            "patch": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Mettre à jour une catégorie spécifique",
                "operationId": "c407bbd00b45b245d6058fcb429a9cca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categorieDestroy{id}": {
            "delete": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Supprimer une catégorie spécifique",
                "operationId": "ee0d5575b127792112f87a41f27989c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commentaires": {
            "get": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Obtenir tous les commentaires",
                "operationId": "ee4d918480e497cee636424744ecfebb",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Commentaire"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commentaireStore{id}": {
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Ajouter un nouveau commentaire à une annonce",
                "operationId": "ed5e237ac5fbc33bff221648f206fa83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'annonce",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "commentaire": {
                                        "type": "string",
                                        "example": "Ceci est un commentaire"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire ajouté avec succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commentaireDestroy{id}": {
            "delete": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Supprimer un commentaire",
                "operationId": "34dca5a7cdff95a9c3c4e63c403d170f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du commentaire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire supprimé avec succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}